Fix fixable warnings.
Commit 8699deaba2c added a list of compiler warnings which
lead to a lot of noise during build.
Non of the issues found seemed to be really relevant, still
this patch attempts to fix all the issues that could be address:
- comparisons between signed and unsigned.
- unreachable code: unreachable return statement.
- shadowing of global symbols (don't use 'index' as a symbol!).
- order of static and const.

Other issues cannot be fixed:
- when functions are inlined some code paths in the function might
  be unreachable. The compiler should deal with this but not warn!
- enums in a switch statement. gcc braindeadlessly complains even
  if there is a default: case.

+++ src/rhd.h
@@ -186,7 +186,7 @@ typedef struct RHDRec {
 
 /* rhd_driver.c */
 /* Some handy functions that makes life so much more readable */
-int RHDReadPCIBios(RHDPtr rhdPtr, unsigned char **prt);
+unsigned int RHDReadPCIBios(RHDPtr rhdPtr, unsigned char **prt);
 CARD32 _RHDRegRead(int scrnIndex, CARD16 offset);
 #define RHDRegRead(ptr, offset) _RHDRegRead((ptr)->scrnIndex, (offset))
 void _RHDRegWrite(int scrnIndex, CARD16 offset, CARD32 value);
@@ -227,7 +227,7 @@ void RHDDebugVerb(int scrnIndex, int verb, const char *format, ...);
 void RHDDebugContVerb(int verb, const char *format, ...);
 #define RHDFUNC(ptr) RHDDebug((ptr)->scrnIndex, "FUNCTION: %s\n", __func__)
 #define RHDFUNCI(scrnIndex) RHDDebug(scrnIndex, "FUNCTION: %s\n", __func__)
-void RhdDebugDump(int scrnIndex, unsigned char *start, unsigned long size);
+void RhdDebugDump(int scrnIndex, unsigned char *start, int size);
 #ifdef RHD_DEBUG
 # define DEBUGP(x) {x;}
 #else
